backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.7.3
pydantic==2.8.2
python-dotenv==1.0.1
selenium==4.23.1
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.27.2
slowapi==0.1.9
